+++ dart/common/Console.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017, The DART development contributors
+ * Copyright (c) 2011-2018, The DART development contributors
  * All rights reserved.
  *
  * The list of contributors can be found at:
@@ -37,16 +37,16 @@
 #include <ostream>
 
 /// \brief Output a message
-#define dtmsg (dart::common::colorMsg("Msg", 32))
+#define dtmsg (::dart::common::colorMsg("Msg", 32))
 
 /// \brief Output a debug message
-#define dtdbg (dart::common::colorMsg("Dbg", 36))
+#define dtdbg (::dart::common::colorMsg("Dbg", 36))
 
 /// \brief Output a warning message
-#define dtwarn (dart::common::colorErr("Warning", __FILE__, __LINE__, 33))
+#define dtwarn (::dart::common::colorErr("Warning", __FILE__, __LINE__, 33))
 
 /// \brief Output an error message
-#define dterr (dart::common::colorErr("Error", __FILE__, __LINE__, 31))
+#define dterr (::dart::common::colorErr("Error", __FILE__, __LINE__, 31))
 
 namespace dart {
 namespace common {