+++ dart/dynamics/Skeleton.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017, The DART development contributors
+ * Copyright (c) 2011-2018, The DART development contributors
  * All rights reserved.
  *
  * The list of contributors can be found at:
@@ -914,6 +914,14 @@ std::vector<const BodyNode*> Skeleton::getBodyNodes(
 }
 
 //==============================================================================
+bool Skeleton::hasBodyNode(const BodyNode* bodyNode) const
+{
+  return std::find(
+      mSkelCache.mBodyNodes.begin(), mSkelCache.mBodyNodes.end(), bodyNode)
+      != mSkelCache.mBodyNodes.end();
+}
+
+//==============================================================================
 template <class ObjectT, std::size_t (ObjectT::*getIndexInSkeleton)() const>
 static std::size_t templatedGetIndexOf(const Skeleton* _skel, const ObjectT* _obj,
                                   const std::string& _type, bool _warning)
@@ -1056,6 +1064,18 @@ std::vector<const Joint*> Skeleton::getJoints(const std::string& name) const
     return {joint};
   else
     return std::vector<const Joint*>();
+}
+
+//==============================================================================
+bool Skeleton::hasJoint(const Joint* joint) const
+{
+  return std::find_if(
+      mSkelCache.mBodyNodes.begin(), mSkelCache.mBodyNodes.end(),
+      [&joint](const BodyNode* bodyNode)
+      {
+        return bodyNode->getParentJoint() == joint;
+      })
+      != mSkelCache.mBodyNodes.end();
 }
 
 //==============================================================================